chore: Moves the datasource components to the components folder
