+++ superset-frontend/src/explore/components/controls/DatasourceControl/index.jsx
@@ -24,8 +24,10 @@ import { t, styled, supersetTheme } from '@superset-ui/core';
 import { Dropdown, Menu } from 'src/common/components';
 import { Tooltip } from 'src/components/Tooltip';
 import Icons from 'src/components/Icons';
-import ChangeDatasourceModal from 'src/datasource/ChangeDatasourceModal';
-import DatasourceModal from 'src/datasource/DatasourceModal';
+import {
+  ChangeDatasourceModal,
+  DatasourceModal,
+} from 'src/components/Datasource';
 import { postForm } from 'src/explore/exploreUtils';
 import Button from 'src/components/Button';
 import ErrorAlert from 'src/components/ErrorMessage/ErrorAlert';
