+++ superset-frontend/src/views/CRUD/data/dataset/DatasetList.tsx
@@ -32,7 +32,7 @@ import {
 import { ColumnObject } from 'src/views/CRUD/data/dataset/types';
 import { useListViewResource } from 'src/views/CRUD/hooks';
 import ConfirmStatusChange from 'src/components/ConfirmStatusChange';
-import DatasourceModal from 'src/datasource/DatasourceModal';
+import { DatasourceModal } from 'src/components/Datasource';
 import DeleteModal from 'src/components/DeleteModal';
 import handleResourceExport from 'src/utils/export';
 import ListView, {
